requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
bitsandbytes==0.48.2
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
configparser==7.2.0
datasets==4.4.1
defusedxml==0.7.1
dill==0.4.0
filelock==3.20.0
fonttools==4.60.1
fpdf2==2.8.5
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
halo==0.0.31
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
interpres==0.4.0b6
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
log-symbols==0.0.14
marisa-trie==1.3.1
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.6
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
pillow==12.0.0
polib==1.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
pyarrow==22.0.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
questionary==2.1.1
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sentencepiece==0.2.1
setuptools==80.9.0
shutils==0.1.0
six==1.17.0
sniffio==1.3.1
spinners==0.0.24
sympy==1.14.0
termcolor==3.2.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
